Cleaned up PolaronHamiltonian
