+++ CoherentState.py
@@ -1,42 +1,7 @@
 import numpy as np
 from polaron_functions import kcos_func
 from scipy.integrate import odeint
-
-
-def amplitude_update(variables_t, t, coherent_state, hamiltonian):
-    # here on can write any method induding Runge-Kutta 4
-
-    [P, aIBi, mI, mB, n0, gBB] = hamiltonian.Params
-
-    # Here I need an original grid
-    dv = coherent_state.grid.dV()
-
-    # Split variables into x and p
-    [x_t, p_t] = np.split(variables_t, 2)
-    PB_t = coherent_state.get_PhononMomentum()
-
-    h_x = 2 * hamiltonian.gnum * np.sqrt(n0) * hamiltonian.Wk_grid +\
-        x_t * (hamiltonian.Omega0_grid - hamiltonian.kcos * (P - PB_t) / mI) +\
-        hamiltonian.gnum * hamiltonian.Wk_grid * np.dot(hamiltonian.Wk_grid, x_t * dv)
-    h_y = p_t * (hamiltonian.Omega0_grid - hamiltonian.kcos * (P - PB_t) / mI) +\
-        hamiltonian.gnum * hamiltonian.Wki_grid * np.dot(hamiltonian.Wki_grid, p_t * dv)
-
-    return np.append(h_y, -1 * h_x)
-
-
-def phase_update(variables_t, t, coherent_state, hamiltonian):
-
-    [P, aIBi, mI, mB, n0, gBB] = hamiltonian.Params
-
-    # Here I need the original grid
-    dv = coherent_state.grid.dV()
-
-    # Split variables into x and p
-    [x_t, p_t] = np.split(coherent_state.amplitude, 2)
-    PB_t = coherent_state.get_PhononMomentum()
-
-    return hamiltonian.gnum * n0 + hamiltonian.gnum * np.sqrt(n0) * np.dot(hamiltonian.Wk_grid, x_t * dv) +\
-        (P**2 - PB_t**2) / (2 * mI)
+from PolaronHamiltonian import amplitude_update, phase_update
 
 
 class CoherentState:
